Check if column user_name exists before removing it.

+++ db/migrate/20161006023701_remove_username_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveUsernameFromUsers < ActiveRecord::Migration[5.0]
   def change
-  	remove_column :users, :user_name
+  	if column_exists? :users, :user_name
+  		remove_column :users, :user_name
+  	end
   end
 end
